+++ R/aliases.R
@@ -13,7 +13,6 @@ NULL
 #' Alias to \link{bs4DashboardPage}
 #' See \link{bs4DashboardPage} for full details
 #'
-#' @inheritParams bs4DashPage
 #' @rdname bs4DashPage
 #'
 #' @examples
@@ -99,7 +98,6 @@ dashboardPage <- bs4DashPage
 #' Alias to \link{bs4DashNavbar}
 #' See \link{bs4DashNavbar} for full details
 #'
-#' @inheritParams bs4DashNavbar
 #' @rdname bs4DashNavbar
 #'
 #' @export
@@ -112,7 +110,6 @@ dashboardHeader <- bs4DashNavbar
 #' Alias to \link{bs4DropdownMenu}
 #' See \link{bs4DropdownMenu} for full details
 #'
-#' @inheritParams bs4DropdownMenu
 #' @rdname bs4DropdownMenu
 #'
 #' @export
@@ -125,7 +122,6 @@ dropdownMenu <- bs4DropdownMenu
 #' Alias to \link{bs4DropdownMenuItem}
 #' See \link{bs4DropdownMenuItem} for full details
 #'
-#' @inheritParams bs4DropdownMenuItem
 #' @rdname bs4DropdownMenuItem
 #'
 #' @export
@@ -138,7 +134,6 @@ dropdownMenuItem <- bs4DropdownMenuItem
 #' Alias to \link{bs4DashControlbar}
 #' See \link{bs4DashControlbar} for full details
 #'
-#' @inheritParams bs4DashControlbar
 #' @rdname bs4DashControlbar
 #'
 #' @export
@@ -189,7 +184,6 @@ updateControlbarMenu <- updatebs4ControlbarMenu
 #' Alias to \link{bs4DashFooter}
 #' See \link{bs4DashFooter} for full details
 #'
-#' @inheritParams bs4DashFooter
 #' @rdname bs4DashFooter
 #'
 #' @export
@@ -202,7 +196,6 @@ dashboardFooter <- bs4DashFooter
 #' Alias to \link{bs4DashSidebar}
 #' See \link{bs4DashSidebar} for full details
 #'
-#' @inheritParams bs4DashSidebar
 #' @rdname bs4DashSidebar
 #'
 #' @export
@@ -215,7 +208,6 @@ dashboardSidebar <- bs4DashSidebar
 #' Alias to \link{bs4SidebarHeader}
 #' See \link{bs4SidebarHeader} for full details
 #'
-#' @inheritParams bs4SidebarHeader
 #' @rdname bs4SidebarHeader
 #'
 #' @export
@@ -228,7 +220,6 @@ sidebarHeader <- bs4SidebarHeader
 #' Alias to \link{bs4SidebarMenu}
 #' See \link{bs4SidebarMenu} for full details
 #'
-#' @inheritParams bs4SidebarMenu
 #' @rdname bs4SidebarMenu
 #'
 #' @export
@@ -241,7 +232,6 @@ sidebarMenu <- bs4SidebarMenu
 #' Alias to \link{bs4SidebarUserPanel}
 #' See \link{bs4SidebarUserPanel} for full details
 #'
-#' @inheritParams bs4SidebarUserPanel
 #' @rdname bs4SidebarUserPanel
 #'
 #' @export
@@ -254,7 +244,6 @@ sidebarUserPanel <- bs4SidebarUserPanel
 #' Alias to \link{bs4SidebarMenuItem}
 #' See \link{bs4SidebarMenuItem} for full details
 #'
-#' @inheritParams bs4SidebarMenuItem
 #' @rdname bs4SidebarMenuItem
 #'
 #' @export
@@ -267,7 +256,6 @@ menuItem <- bs4SidebarMenuItem
 #' Alias to \link{bs4SidebarMenuSubItem}
 #' See \link{bs4SidebarMenuSubItem} for full details
 #'
-#' @inheritParams bs4SidebarMenuSubItem
 #' @rdname bs4SidebarMenuSubItem
 #'
 #' @export
@@ -280,7 +268,6 @@ menuSubItem <- bs4SidebarMenuSubItem
 #' Alias to \link{bs4DashBody}
 #' See \link{bs4DashBody} for full details
 #'
-#' @inheritParams bs4DashBody
 #' @rdname bs4DashBody
 #'
 #' @export
@@ -291,7 +278,6 @@ dashboardBody <- bs4DashBody
 #' Alias to \link{bs4TabItems}
 #' See \link{bs4TabItems} for full details
 #'
-#' @inheritParams bs4TabItems
 #' @rdname bs4TabItems
 #'
 #' @export
@@ -303,7 +289,6 @@ tabItems <- bs4TabItems
 #' Alias to \link{bs4TabItem}
 #' See \link{bs4TabItem} for full details
 #'
-#' @inheritParams bs4TabItem
 #' @rdname bs4TabItem
 #'
 #' @export
@@ -317,7 +302,6 @@ tabItem <- bs4TabItem
 #' Alias to \link{bs4Card}
 #' See \link{bs4Card} for full details
 #'
-#' @inheritParams bs4Card
 #' @rdname bs4Card
 #'
 #' @export
@@ -330,7 +314,6 @@ box <- bs4Card
 #' Alias to \link{bs4TabCard}
 #' See \link{bs4TabCard} for full details
 #'
-#' @inheritParams bs4TabCard
 #' @rdname bs4TabCard
 #'
 #' @export
@@ -342,7 +325,6 @@ tabBox <- bs4TabCard
 #' Alias to \link{bs4TabSetPanel}
 #' See \link{bs4TabSetPanel} for full details
 #'
-#' @inheritParams bs4TabSetPanel
 #' @rdname bs4TabSetPanel
 #'
 #' @export
@@ -356,7 +338,6 @@ tabsetPanel <- bs4TabSetPanel
 #' Alias to \link{bs4TabPanel}
 #' See \link{bs4TabPanel} for full details
 #'
-#' @inheritParams bs4TabPanel
 #' @rdname bs4TabPanel
 #'
 #' @export
@@ -370,7 +351,6 @@ tabPanel <- bs4TabPanel
 #' Alias to \link{bs4InfoBox}
 #' See \link{bs4InfoBox} for full details
 #'
-#' @inheritParams bs4InfoBox
 #' @rdname bs4InfoBox
 #'
 #' @export
@@ -382,7 +362,7 @@ infoBox <- bs4InfoBox
 
 #' Alias to \link{bs4InfoBoxOutput}
 #' See \link{bs4InfoBoxOutput} for full details
-#'
+#' 
 #' @inheritParams bs4InfoBoxOutput
 #' @rdname bs4InfoBoxOutput
 #'
@@ -409,7 +389,6 @@ renderInfoBox <- renderbs4InfoBox
 #' Alias to \link{bs4ValueBox}
 #' See \link{bs4ValueBox} for full details
 #'
-#' @inheritParams bs4ValueBox
 #' @rdname bs4ValueBox
 #'
 #' @export
@@ -422,7 +401,6 @@ valueBox <- bs4ValueBox
 #' Alias to \link{bs4ValueBoxOutput}
 #' See \link{bs4ValueBoxOutput} for full details
 #'
-#' @inheritParams bs4ValueBoxOutput
 #' @rdname bs4ValueBoxOutput
 #'
 #' @export
@@ -435,7 +413,6 @@ valueBoxOutput <- bs4ValueBoxOutput
 #' Alias to \link{renderbs4ValueBox}
 #' See \link{renderbs4ValueBox} for full details
 #'
-#' @inheritParams renderbs4ValueBox
 #' @rdname renderbs4ValueBox
 #'
 #' @export
@@ -448,7 +425,6 @@ renderValueBox <- renderbs4ValueBox
 #' Alias to \link{updatebs4TabSetPanel}
 #' See \link{updatebs4TabSetPanel} for full details
 #'
-#' @inheritParams updatebs4TabSetPanel
 #' @rdname updatebs4TabSetPanel
 #'
 #' @export
@@ -460,8 +436,30 @@ updateTabsetPanel <- updatebs4TabSetPanel
 #' Alias to \link{ updatebs4TabItems}
 #' See \link{updatebs4TabItems} for full details
 #'
-#' @inheritParams  updatebs4TabItems
 #' @rdname  updatebs4TabItems
 #'
 #' @export
 updateTabItems <- updatebs4TabItems
+
+
+
+
+#' Alias to \link{bs4CardSidebar}
+#' See \link{bs4CardSidebar} for full details
+#'
+#' @rdname bs4CardSidebar
+#'
+#' @export
+cardSidebar <- bs4CardSidebar
+
+
+
+
+
+#' Alias to \link{updatebs4CardSidebar}
+#' See \link{updatebs4CardSidebar} for full details
+#'
+#' @rdname updatebs4CardSidebar
+#'
+#' @export
+updateCardSidebar <- updatebs4CardSidebar