+++ NEWS.md
@@ -1,5 +1,32 @@
 # bs4Dash 0.6.0.9000
+
+## Breaking changes
+- `updatebs4TabSetPanel()`: selected takes the name of the tab instead of its index. This is more convenient (same as in shinydashboard)
+- remove all sidebar related parameters from `bs4Card()`. This is now part of
+ the new `bs4CardSidebar()`
+- remove dropdownIcon parameter from `bs4Card()`. It is now part of the `dropdownItemList()`
+- remove labelStatus, labelText and labelTooltip params from `bs4Card()`. This is to reduce the number of parameters of cards
+- By default, closable is now FALSE for `bs4Card()`
+- value is now mandatory in `bs4InfoBox()`
+
+## New features
+- new `bs4CardLabel()` to add text labels in `bs4Card()`
+- new `bs4CardSidebar`: access the status via input$id. Add `updatebs4CardSidebar()` to toggle the card sidebar
+- add `bs4PopoverServer()`, `bs4PopoverUI()`, `bs4TooltipUI()` and `bs4TooltipServer()`
+- `bs4TabSetPanel()`: you may access the currently selected tab with input$id
+- add type parameter to `bs4TabSetPanel()` and `bs4TabCard()` so as to select
+between pills or tabs (cosmetic choice). If type is not provided, the behaviour is unchanged (pills are the default)
+- add fixed param to `bs4DashFooter()`
+- add fixed parameter to `bs4DashSidebar()`. Thanks @mppallante
+- re-add fixed parameter to `bs4DashNavbar()`. This is reintroduced by adminlte with a new css class name
 - add animated, label parameters to `bs4ProgressBar()`
+- add iconStatus parameter to `bs4InfoBox()` to allow icon color
+to be independant from the card status (https://adminlte.io/themes/v3/pages/widgets.html). Thanks @rolando-gsw
+
+## Bug Fix
+- fix #86 : cannot use `updatebs4TabSetPanel()` with `renderUI`. Thanks @jyjek
+- add condition argument to `bs4SidebarMenuItem()`. See https://github.com/RinteRface/bs4Dash/issues/35. Thanks @aldomann
+- some vignettes examples were not up to date
 
 # bs4Dash 0.5.0